Added new token filters for Japanese sutegana (捨て仮名)
Cherry-picked from: https://github.com/apache/lucene/pull/12915 by @daixque

Context

Sutegana (捨て仮名) is small letter of hiragana and katakana in Japanese. In the old Japanese text, sutegana (捨て仮名) is not used unlike in the modern texts. For example:

"ストップウォッチ" is written as "ストツプウオツチ"
"ちょっとまって" is written as "ちよつとまつて"

So it's meaningful to normalize Sutegana to normal (uppercase) characters if we search against the corpuses which includes old Japanese texts such as patents, legal documents, contract policies, etc.